requirements.txt
branca
folium==0.19.2
geopandas==1.0.1
matplotlib
numpy>=1.23.0
scikit-image>=0.19.0
pandas
Pillow
plotly
Requests==2.32.3
scipy==1.15.2
seaborn
sentinelhub==3.11.1
streamlit==1.41.1
streamlit_folium==0.24.0